#5132 but easier to review
